backend-python/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
motor